requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.4
astroid==2.5.6
cffi==1.14.5
Django==3.2
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
flake8==3.9.2
flake8-html==0.4.1
importlib-metadata==4.0.1
isort==5.8.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.9.0
PyJWT==1.7.1
pylint==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
zipp==3.4.1
